Use the same types for the two values in assertEqual()

+++ tests/unit/dom/test1.cpp
@@ -10,6 +10,6 @@ void webMain()
 	//Test JS Arrays
 	client::Array* ar = new client::Array(1,2,3);
 
-	assertEqual(*(*ar)[0], 1, "JS Array 1/2");
-	assertEqual(*(*ar)[1], 2, "JS Array 2/2");
+	assertEqual((int)*(*ar)[0], 1, "JS Array 1/2");
+	assertEqual((int)*(*ar)[1], 2, "JS Array 2/2");
 }
